+++ conveyor/interfaces/slot/slot_element_helper_mock.go
@@ -47,7 +47,7 @@ type SlotElementHelperMock struct {
 	GetPayloadPreCounter uint64
 	GetPayloadMock       mSlotElementHelperMockGetPayload
 
-	GetStateFunc       func() (r uint16)
+	GetStateFunc       func() (r int)
 	GetStateCounter    uint64
 	GetStatePreCounter uint64
 	GetStateMock       mSlotElementHelperMockGetState
@@ -906,7 +906,7 @@ type SlotElementHelperMockGetStateExpectation struct {
 }
 
 type SlotElementHelperMockGetStateResult struct {
-	r uint16
+	r int
 }
 
 //Expect specifies that invocation of SlotElementHelper.GetState is expected from 1 to Infinity times
@@ -922,7 +922,7 @@ func (m *mSlotElementHelperMockGetState) Expect() *mSlotElementHelperMockGetStat
 }
 
 //Return specifies results of invocation of SlotElementHelper.GetState
-func (m *mSlotElementHelperMockGetState) Return(r uint16) *SlotElementHelperMock {
+func (m *mSlotElementHelperMockGetState) Return(r int) *SlotElementHelperMock {
 	m.mock.GetStateFunc = nil
 	m.expectationSeries = nil
 
@@ -944,12 +944,12 @@ func (m *mSlotElementHelperMockGetState) ExpectOnce() *SlotElementHelperMockGetS
 	return expectation
 }
 
-func (e *SlotElementHelperMockGetStateExpectation) Return(r uint16) {
+func (e *SlotElementHelperMockGetStateExpectation) Return(r int) {
 	e.result = &SlotElementHelperMockGetStateResult{r}
 }
 
 //Set uses given function f as a mock of SlotElementHelper.GetState method
-func (m *mSlotElementHelperMockGetState) Set(f func() (r uint16)) *SlotElementHelperMock {
+func (m *mSlotElementHelperMockGetState) Set(f func() (r int)) *SlotElementHelperMock {
 	m.mainExpectation = nil
 	m.expectationSeries = nil
 
@@ -958,7 +958,7 @@ func (m *mSlotElementHelperMockGetState) Set(f func() (r uint16)) *SlotElementHe
 }
 
 //GetState implements github.com/insolar/insolar/conveyor/interfaces/slot.SlotElementHelper interface
-func (m *SlotElementHelperMock) GetState() (r uint16) {
+func (m *SlotElementHelperMock) GetState() (r int) {
 	counter := atomic.AddUint64(&m.GetStatePreCounter, 1)
 	defer atomic.AddUint64(&m.GetStateCounter, 1)
