+++ conveyor/slot_element.go
@@ -39,7 +39,7 @@ type slotElement struct {
 	payload          interface{} // nolint
 	postponedError   error
 	stateMachineType statemachine.StateMachineType
-	state            uint16
+	state            int
 
 	nextElement      *slotElement
 	activationStatus ActivationStatus
@@ -95,7 +95,7 @@ func (se *slotElement) GetType() int {
 }
 
 // GetState implements SlotElementReadOnly
-func (se *slotElement) GetState() uint16 {
+func (se *slotElement) GetState() int {
 	return se.state
 }
